test(notifications): isolate delivery preferences
